cedarling::jwt_config
Struct JwtConfig

pub struct JwtConfig {
    pub jwks: Option<String>,
    pub jwt_sig_validation: bool,
    pub jwt_status_validation: bool,
    pub signature_algorithms_supported: HashSet<Algorithm>,
}

The set of Bootstrap properties related to JWT validation.

Fields
jwks: Option<String>

A Json Web Key Store (JWKS) with public keys.

If this is used, Cedarling will no longer try to fetch JWK Stores from a trustede identity provider and stick to using the local JWKS.

jwt_sig_validation: bool

Check the signature for all the Json Web Tokens.

This Requires the iss claim to be present in all the tokens and and the scheme must be https.

This setting overrides the iss validation settings in the following:

access_token_config

id_token_config

userinfo_token_config

jwt_status_validation: bool

Whether to check the status of the JWT.

On startup, the Cedarling will fetch and retreive the latest Status List JWT from the .well-known/openid-configuration via the status_list_endpoint claim and cache it.

See the IETF Draft for more info.

signature_algorithms_supported: HashSet<Algorithm>

Only tokens signed with algorithms in this list can be valid.

Implementations

impl JwtConfig

pub fn new_without_validation() -> Self

Creates a new JwtConfig instance with validation turned off for all tokens.

pub fn allow_all_algorithms(self) -> Self

Adds all supported algorithms to to signature_algorithms_supported.