cedarling::jwt_config

Struct TokenValidationConfig

pub struct TokenValidationConfig {
    pub iss_validation: bool,
    pub aud_validation: bool,
    pub sub_validation: bool,
    pub jti_validation: bool,
    pub iat_validation: bool,
    pub exp_validation: bool,
    pub nbf_validation: bool,
}

Validation options related to JSON Web Tokens (JWT).

This struct provides the configuration for validating common JWT claims (iss, aud, sub, jti, exp, nbf) across different types of JWTs.

The default configuration for Access Tokens, ID Tokens, and Userinfo Tokens can be easily instantiated via the provided methods.

Fields
iss_validation: bool

Requires the iss claim to be present in the JWT and the scheme must be https.

aud_validation: bool

Requires the aud claim to be present in the JWT.

sub_validation: bool

Requires the sub claim to be present in the JWT.

jti_validation: bool

Requires the jti claim to be present in the JWT.

iat_validation: bool

Requires the iat claim to be present in the JWT.

exp_validation: bool

Requires the exp claim to be present in the JWT and the current timestamp isn’t past the specified timestamp in the token.

nbf_validation: bool

Requires the nbf claim to be present in the JWT.

Implementations

impl TokenValidationConfig

pub fn required_claims(&self) -> HashSet<Box<str>>

Collects all the required claims into a HashSet.

pub fn access_token() -> Self

Returns a default configuration for validating Access Tokens.

This configuration requires the following:

iss (Issuer)

jti (JWT ID)

exp (Expiration)

pub fn id_token() -> Self

Returns a default configuration for validating ID Tokens.

This configuration requires the following:

iss (Issuer)

aud (Audience)

sub (Subject)

exp (Expiration)

pub fn userinfo_token() -> Self

Returns a default configuration for validating Userinfo Tokens.

This configuration requires the following:

iss (issuer)

aud (audience)

sub (subject)

exp (expiration)