cedarling

Module log_config

Log config module

Structs

LockServiceConfig

Config for the lock logger that are set using the bootstrap configs.

LogConfig

A set of properties used to configure logging in the Cedarling application.

MemoryLogConfig

Configuration for memory log.

MemoryLogConfigRaw

Raw memory log config

Enums

LogTypeConfig

Log type configuration. CEDARLING_LOG_TYPE in bootstrap properties documentation. Current type represent this value.

StdOutLoggerMode

Mode for stdout logger in native builds. Supports both immediate and asynchronous logging.

StdOutMode

Mode for stdout logging. Is used in BootstrapConfigRaw