cedarling
Struct BootstrapConfigRaw

pub struct BootstrapConfigRaw {

    pub application_name: String,
    pub policy_store_uri: Option<String>,
    pub policy_store_id: String,
    pub log_type: LoggerType,
    pub log_level: LogLevel,
    pub log_ttl: Option<u64>,
    pub log_max_items: Option<usize>,
    pub log_max_item_size: Option<usize>,
    pub decision_log_user_claims: Vec<String>,
    pub decision_log_workload_claims: Vec<String>,
    pub decision_log_default_jwt_id: String,
    pub user_authz: FeatureToggle,
    pub workload_authz: FeatureToggle,
    pub principal_bool_operation: JsonRule,
    pub mapping_iss: Option<String>,
    pub mapping_user: Option<String>,
    pub mapping_workload: Option<String>,
    pub mapping_role: Option<String>,
    pub unsigned_role_id_src: UnsignedRoleIdSrc,
    pub local_jwks: Option<String>,
    pub local_policy_store: Option<String>,
    pub policy_store_local_fn: Option<String>,
    pub max_default_entities: Option<usize>,
    pub max_base64_size: Option<usize>,
    pub jwt_sig_validation: FeatureToggle,
    pub jwt_status_validation: FeatureToggle,
    pub jwt_signature_algorithms_supported: HashSet<Algorithm>,
    pub id_token_trust_mode: IdTokenTrustMode,
    pub lock: FeatureToggle,
    pub lock_server_configuration_uri: Option<String>,
    pub dynamic_configuration: FeatureToggle,
    pub lock_ssa_jwt: Option<String>,
    pub audit_log_interval: u64,
    pub audit_health_interval: u64,
    pub audit_telemetry_interval: u64,
    pub listen_sse: FeatureToggle,
    pub accept_invalid_certs: FeatureToggle,
    pub token_cache_max_ttl: usize,

}

Struct that represent mapping mapping Bootstrap properties to be JSON and YAML compatible from link

This structure is used to deserialize values from ENV VARS so json keys is same as keys in environment variables

Fields
application_name: String

Human friendly identifier for the application

policy_store_uri: Option<String>

Location of policy store JSON, used if policy store is not local, or retreived from Lock Master.

policy_store_id: String

An identifier for the policy store.

log_type: LoggerType

How the Logs will be presented.

log_level: LogLevel

Log level filter for logging. TRACE is lowest. FATAL is highest.

log_ttl: Option<u64>

If log_type is set to [LogType::Memory], this is the TTL (time to live) of log entities in seconds.

log_max_items: Option<usize>

Maximum number of log entities that can be stored using [LogType::Memory]. If value is 0, there is no limit. But if None, default value is applied.

log_max_item_size: Option<usize>

Maximum size of a single log entity in bytes using [LogType::Memory]. If value is 0, there is no limit. But if None, default value is applied.

decision_log_user_claims: Vec<String>

List of claims to map from user entity, such as [“sub”, “email”, “username”, …]

decision_log_workload_claims: Vec<String>

List of claims to map from user entity, such as [“client_id”, “rp_id”, …]

decision_log_default_jwt_id: String

Token claims that will be used for decision logging. Default is jti, but perhaps some other claim is needed.

user_authz: FeatureToggle

When enabled, Cedar engine authorization is queried for a User principal.

workload_authz: FeatureToggle

When enabled, Cedar engine authorization is queried for a Workload principal.

principal_bool_operation: JsonRule

Specifies what boolean operation to use for the USER and WORKLOAD when making authz (authorization) decisions.

Use JsonLogic.

Default value:

{
    "and" : [
        {"===": [{"var": "Jans::Workload"}, "ALLOW"]},
        {"===": [{"var": "Jans::User"}, "ALLOW"]}
    ]
}

mapping_iss: Option<String>

Mapping name of cedar schema TrustedIssuer entity

mapping_user: Option<String>

Mapping name of cedar schema User entity

mapping_workload: Option<String>

Mapping name of cedar schema Workload entity.

mapping_role: Option<String>

Mapping name of cedar schema Role entity.

unsigned_role_id_src: UnsignedRoleIdSrc

Mapping name of cedar schema Role entity.

local_jwks: Option<String>

Path to a local file pointing containing a JWKS.

local_policy_store: Option<String>

JSON object with policy store

policy_store_local_fn: Option<String>

Path to a Policy Store JSON file

max_default_entities: Option<usize>

Maximum number of default entities allowed in a policy store. This prevents DoS attacks by limiting the number of entities that can be loaded. If value is 0, there is no limit. But if None, default value is applied.

max_base64_size: Option<usize>

Maximum size of base64-encoded default entity strings in bytes. This prevents memory exhaustion attacks from extremely large base64 strings. If value is 0, there is no limit. But if None, default value is applied.

jwt_sig_validation: FeatureToggle

Whether to check the signature of all JWT tokens.

This requires that an iss (Issuer) claim is present on each token.

jwt_status_validation: FeatureToggle

Whether to check the status of the JWT. On startup.

Cedarling will fetch and retreive the latest Status List JWT from the .well-known/openid-configuration via the status_list_endpoint claim and cache it. See the IETF Draft for more info.

jwt_signature_algorithms_supported: HashSet<Algorithm>

Cedarling will only accept tokens signed with these algorithms.

id_token_trust_mode: IdTokenTrustMode

Varying levels of validations based on the preference of the developer.

Strict Mode

Strict mode requires: 1. id_token aud matches the access_token client_id; 2. if a Userinfo token is present, the sub matches the id_token, and that the aud matches the access token client_id.

lock: FeatureToggle

If Enabled, the Cedarling will connect to the Lock Master for policies, and subscribe for SSE events.

lock_server_configuration_uri: Option<String>

URI where Cedarling can get JSON file with all required metadata about Lock Master, i.e. .well-known/lock-master-configuration.

Required if LOCK == enabled.

dynamic_configuration: FeatureToggle

Controls whether Cedarling should listen for SSE config updates.

lock_ssa_jwt: Option<String>

SSA for DCR in a Lock Master deployment. The Cedarling will validate this SSA JWT prior to DCR.

audit_log_interval: u64

How often to send log messages to Lock Master (0 to turn off trasmission).

audit_health_interval: u64

How often to send health messages to Lock Master (0 to turn off transmission).

audit_telemetry_interval: u64

How often to send telemetry messages to Lock Master (0 to turn off transmission).

listen_sse: FeatureToggle

Controls whether Cedarling should listen for updates from the Lock Server.

accept_invalid_certs: FeatureToggle

Allow interaction with a Lock server with invalid certificates. Used for testing.

token_cache_max_ttl: usize

Allows to limit maximum token cache TTL in seconds. Zero means no token cache TTL limit.

Implementations

impl BootstrapConfigRaw

pub fn from_raw_config_and_env( raw: Option<BootstrapConfigRaw>, ) -> Result<Self, BootstrapConfigLoadingError>

Construct BootstrapConfig from environment variables and BootstrapConfigRaw config. Environment variables have bigger priority.