cedarling

Trait TrustedIssuerLoadingInfo

pub trait TrustedIssuerLoadingInfo {
    // Required methods
    fn is_trusted_issuer_loaded_by_name(&self, issuer_id: &str) -> bool;
    fn is_trusted_issuer_loaded_by_iss(&self, iss_claim: &str) -> bool;
    fn total_issuers(&self) -> usize;
    fn loaded_trusted_issuers_count(&self) -> usize;
    fn loaded_trusted_issuer_ids(&self) -> HashSet<String>;
    fn failed_trusted_issuer_ids(&self) -> HashSet<String>;
}

Provides information about the loading status of trusted issuers.

This trait is implemented by services that can report on the progress of loading trusted issuers, including which issuers are loaded, which have failed, and overall loading statistics.

Required Methods

fn is_trusted_issuer_loaded_by_name(&self, issuer_id: &str) -> bool

Returns true if the trusted issuer with the given ID (policy store key) is loaded.

fn is_trusted_issuer_loaded_by_iss(&self, iss_claim: &str) -> bool

Returns true if the trusted issuer with the given iss claim is loaded.

fn total_issuers(&self) -> usize

Returns the total number of trusted issuers that are expected to be loaded.

fn loaded_trusted_issuers_count(&self) -> usize

Returns the number of trusted issuers that have been loaded.

fn loaded_trusted_issuer_ids(&self) -> HashSet<String>

Returns all issuer IDs in the index of loaded trusted issuers.

This returns a set of unique issuer IDs that have been successfully loaded.

fn failed_trusted_issuer_ids(&self) -> HashSet<String>

Returns issuer IDs that failed to load.

This returns a set of unique issuer IDs that encountered errors during loading. Failed issuers are still counted toward the total processing count.