Crate cedarling

Cedarling

The Cedarling is a performant local authorization service that runs the Rust Cedar Engine. Cedar policies and schema are loaded at startup from a locally cached “Policy Store”. In simple terms, the Cedarling returns the answer: should the application allow this action on this resource given these JWT tokens. “Fit for purpose” policies help developers build a better user experience. For example, why display form fields that a user is not authorized to see? The Cedarling is a more productive and flexible way to handle authorization.

Modules

authorization_config

Authorization config module

entity_builder_config

Entity builder config module

jwt_config

JWT config module

lock_config

Lock config module

log_config

Log config module

policy_store_config

Policy store config module

raw_config

Raw config module

Structs

AuthorizationConfig

Configuration to specify authorization workflow.

AuthorizationConfigRaw

Raw authorization config

AuthorizeResult

Result of authorization and evaluation cedar policy based on the Request and policy store

BootstrapConfig

Bootstrap configuration properties for configuration Cedarling application. link to the documentation.

BootstrapConfigRaw

Struct that represent mapping mapping Bootstrap properties to be JSON and YAML compatible from link

CedarEntityMapping

Cedar entity mapping information

Cedarling

The instance of the Cedarling application. It is safe to share between threads.

EntityBuilderConfig

Bootstrap Configurations for the JWT to Cedar entity mappings

EntityBuilderConfigRaw

Raw entity builder config

EntityData

Cedar policy entity data fields represent EntityUid

EntityNames

Entity names

JsonRule

JsonLogic rule using JsonLogic Default implementation:

JwtConfig

The set of Bootstrap properties related to JWT validation.

JwtConfigRaw

Raw JWT config

LockServiceConfig

Lock service config

LockServiceConfigRaw

Raw lock service config

LogConfig

A set of properties used to configure logging in the Cedarling application.

LogConfigRaw

Raw log config

MemoryLogConfig

Configuration for memory log.

PolicyStoreConfig

PolicyStoreConfig - Configuration for the policy store.

PolicyStoreConfigRaw

Raw policy store config

Request

Box to store authorization data

RequestUnsigned

Box to store authorization data, with any additional principals

UnsignedRoleIdSrc

Unsigned role ID source

Enums

AuthorizeError

Error type for Authorization Service

BootstrapConfigLoadingError

Represents errors that may occur while loading a BootstrapConfig from a file.

FeatureToggle

Enum varians that represent if feature is enabled or disabled

IdTokenTrustMode

ID token trust mode

InitCedarlingError

Errors that can occur during initialization Cedarling.

LogLevel

Log levels Fatal level is the highest, trace is lowest

LogTypeConfig

Log type configuration. CEDARLING_LOG_TYPE in bootstrap properties documentation. Current type represent this value.

PolicyStoreSource

PolicyStoreSource represents the source from which policies will be retrieved.

Traits

LogStorage

Log Storage interface for getting log entries from the storage