cedarling
Module log_config

Log config module

Structs

LockServiceConfig

Config for the lock logger that are set using the bootstrap configs.

LogConfig

A set of properties used to configure logging in the Cedarling application.

LogConfigRaw

Raw log config

MemoryLogConfig

Configuration for memory log.

MemoryLogConfigRaw

Raw memory log config

Enums

LogTypeConfig

Log type configuration. CEDARLING_LOG_TYPE in bootstrap properties documentation. Current type represent this value.