cedarling
Module raw_config

Raw config module

Modules

token_settings

Token settings module

Structs

BootstrapConfigRaw

Struct that represent mapping mapping Bootstrap properties to be JSON and YAML compatible from link

ParseFeatureToggleError

Error parsing feature toggle

ParseLoggerTypeError

Error parsing logger type

ParseTrustModeError

Error parsing trust mode

ParseWorkloadBoolOpError

Error parsing workload bool op

Enums

FeatureToggle

Enum varians that represent if feature is enabled or disabled

LoggerType

Type of logger